Expand Github workflow to be selective about tagging
This changes the Github workflow so when pushing git tags it only triggers only
on tags of the form "vX.Y.Z", and only when such a tag matches the version
number in package.json will an NPM repo push result.

This makes the workflow resilient to simultaneous pushes of multiple tags,
e.g. v1, v1.0, v1.0.0 -- only the latter will do anything.